# docs/api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Replate AI – REST API",
    "version": "1.0.0",
    "description": "Complete REST API for Replate AI, an AI-powered leftover recipe and grocery management application. All protected endpoints require a Firebase ID token supplied as a Bearer token in the Authorization header.",
    "contact": {
      "name": "CSYE7200 Group 03"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5050",
      "description": "Local development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firebase ID Token obtained from /auth/signin"
      }
    },
    "schemas": {
...
      "RecipeIngredient": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "chicken breast"
          },
          "amount": {
            "type": "number",
            "example": 2
          },
          "unit": {
            "type": "string",
            "example": "pieces"
          }
        }
      },
      "Recipe": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "716429"
          },
          "title": {
            "type": "string",
            "example": "Pasta with Garlic"
          },
          "image": {
            "type": "string",
            "format": "uri"
          },
          "readyInMinutes": {
            "type": "integer",
            "example": 30
          },
          "servings": {
            "type": "integer",
            "example": 4
          },
          "summary": {
            "type": "string"
          },
          "sourceUrl": {
            "type": "string",
            "format": "uri"
          },
          "ingredients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeIngredient"
            }
          },
          "instructions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "InventoryItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "item-uuid"
          },
          "ingredientName": {
            "type": "string",
            "example": "chicken"
          },
          "quant": {
            "type": "number",
            "example": 2
          },
          "unit": {
            "type": "string",
            "example": "kg"
          },
          "expiryDate": {
            "type": "string",
            "format": "date",
            "example": "2026-04-01"
          },
          "isExpired": {
            "type": "boolean",
            "example": false
          },
          "s3Url": {
            "type": "string",
            "format": "uri"
          }
        }
      },
      "GroceryListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "item-uuid"
          },
          "name": {
            "type": "string",
            "example": "flour"
          },
          "amount": {
            "type": "number",
            "example": 2
          },
          "unit": {
            "type": "string",
            "example": "cups"
          },
          "isAvailableAtHome": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "GroceryList": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "list-uuid"
          },
          "recipeId": {
            "type": "string",
            "example": "716429"
          },
          "recipeTitle": {
            "type": "string",
            "example": "Pasta with Garlic"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroceryListItem"
            }
          }
        }
      },
      "DietaryPreferences": {
        "type": "object",
        "properties": {
          "restrictions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "vegetarian"
            ]
          },
          "allergies": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "peanuts"
            ]
          },
          "skillLevel": {
            "type": "string",
            "example": "beginner"
          },
          "maxCookingTime": {
            "type": "integer",
            "example": 30
          }
        }
      },
      "UserProfile": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "string",
            "example": "abc123uid"
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "displayName": {
            "type": "string",
            "example": "Jane Doe"
          }
        }
      },
      "NotificationSettings": {
        "type": "object",
        "properties": {
          "expiryRemindersEnabled": {
            "type": "boolean",
            "example": true
          },
          "reminderLeadDays": {
            "type": "integer",
            "example": 2
          },
          "reminderTime": {
            "type": "string",
            "example": "09:00"
          }
        }
      },
      "AppSettings": {
        "type": "object",
        "properties": {
          "themeMode": {
            "type": "string",
            "enum": [
              "light",
              "dark",
              "system"
            ],
            "example": "system"
          },
          "notifications": {
            "$ref": "#/components/schemas/NotificationSettings"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/grocery-list": {
      "post": {
        "summary": "Create a grocery list from a single recipe's missing ingredients",
        "tags": [
          "Grocery List"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "recipeId",
                  "recipeTitle",
                  "missingIngredients"
                ],
                "properties": {
                  "recipeId": {
                    "type": "string",
                    "example": "716429"
                  },
                  "recipeTitle": {
                    "type": "string",
                    "example": "Pasta with Garlic"
                  },
                  "missingIngredients": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/RecipeIngredient"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Grocery list created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "list": {
                      "$ref": "#/components/schemas/GroceryList"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields"
          }
        }
      },
      "get": {
        "summary": "Get all grocery lists for the authenticated user",
        "tags": [
          "Grocery List"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lists returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "lists": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/GroceryList"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/inventory": {
      "get": {
        "summary": "Get the authenticated user's ingredient inventory",
        "tags": [
          "Inventory"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Inventory retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/InventoryItem"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/profile": {
      "get": {
        "summary": "Get the authenticated user's profile and dietary preferences",
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "profile": {
                      "$ref": "#/components/schemas/UserProfile"
                    },
                    "dietaryPreferences": {
                      "$ref": "#/components/schemas/DietaryPreferences"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "patch": {
        "summary": "Update profile fields and/or dietary preferences",
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "displayName": {
                    "type": "string",
                    "example": "Jane Doe"
                  },
                  "dietaryPreferences": {
                    "$ref": "#/components/schemas/DietaryPreferences"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "profile": {
                      "$ref": "#/components/schemas/UserProfile"
                    },
                    "dietaryPreferences": {
                      "$ref": "#/components/schemas/DietaryPreferences"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/recipe-history": {
      "post": {
        "summary": "Save a recipe to the user's history",
        "description": "Silently skips if the recipe is already saved (dedup by recipeId).",
        "tags": [
          "Recipe History"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "recipe"
                ],
                "properties": {
                  "recipe": {
                    "$ref": "#/components/schemas/Recipe"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Recipe saved (or already existed)"
          },
          "400": {
            "description": "Missing recipe or recipe.id"
          }
        }
      },
      "get": {
        "summary": "Get all saved recipes for the authenticated user",
        "description": "Returns recipes ordered by savedAt descending.",
        "tags": [
          "Recipe History"
        ],
        "responses": {
          "200": {
            "description": "Recipe history returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "recipes": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Recipe"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/settings": {
      "get": {
        "summary": "Get the authenticated user's application settings",
        "tags": [
          "Settings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Settings returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "appSettings": {
                      "$ref": "#/components/schemas/AppSettings"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "patch": {
        "summary": "Update application settings",
        "tags": [
          "Settings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "appSettings"
                ],
                "properties": {
                  "appSettings": {
                    "$ref": "#/components/schemas/AppSettings"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Settings updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "appSettings": {
                      "$ref": "#/components/schemas/AppSettings"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  }
}